articles' editors are containers' editors

+++ test/meta_process_test.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from oc_meta.lib.file_manager import get_data
 from oc_meta.run.meta_process import MetaProcess, run_meta_process
 from test.curator_test import reset_server
@@ -10,10 +11,14 @@ import unittest
 
 BASE_DIR = os.path.join('test', 'meta_process')
 
-def delete_output_zip(base_dir:str) -> None:
+def delete_output_zip(base_dir:str, start_time:datetime) -> None:
     for file in os.listdir(base_dir):
         if file.startswith('meta_output') and file.endswith('.zip'):
-            os.remove(os.path.join(base_dir, file))
+            file_creation_time = file.split('meta_output')[1].replace('.zip', '')
+            file_creation_time = datetime.strptime(file_creation_time, '%Y-%m-%dT%H_%M_%S_%f')
+            was_created_after_time = True if file_creation_time > start_time else False
+            if was_created_after_time:
+                os.remove(os.path.join(base_dir, file))
 
 class test_ProcessTest(unittest.TestCase):
     def test_get_data(self):
@@ -26,16 +31,12 @@ class test_ProcessTest(unittest.TestCase):
         reset_server()
         output_folder = os.path.join(BASE_DIR, 'output_1')
         meta_process = MetaProcess(config=os.path.join(BASE_DIR, 'meta_config_1.yaml'))
+        now = datetime.now()
         run_meta_process(meta_process)
         output = list()
         for dirpath, _, filenames in os.walk(os.path.join(output_folder, 'csv')):
             for file in filenames:
                 output.extend(get_data(os.path.join(dirpath, file)))
-        expected_output = [
-            {'id': 'doi:10.17117/na.2015.08.1067 meta:br/0601', 'title': '', 'author': '', 'pub_date': '', 'venue': 'Scientometrics [issn:0138-9130 issn:1588-2861 meta:br/0603]', 'volume': '26', 'issue': '', 'page': '', 'type': 'journal article', 'publisher': 'Consulting Company Ucom [crossref:6623 meta:ra/0601]', 'editor': 'Naimi, Elmehdi [orcid:0000-0002-4126-8519 meta:ra/0602]'}, 
-            {'id': 'issn:1524-4539 issn:0009-7322 meta:br/0602', 'title': 'Circulation', 'author': '', 'pub_date': '', 'venue': '', 'volume': '', 'issue': '', 'page': '', 'type': 'journal', 'publisher': '', 'editor': ''}, 
-            {'id': 'doi:10.9799/ksfan.2012.25.1.069 meta:br/0601', 'title': 'Nonthermal Sterilization And Shelf-life Extension Of Seafood Products By Intense Pulsed Light Treatment', 'author': 'Cheigh, Chan-Ick [orcid:0000-0003-2542-5788 meta:ra/0601]; Mun, Ji-Hye [meta:ra/0602]; Chung, Myong-Soo [meta:ra/0603]', 'pub_date': '2012-03-31', 'venue': 'The Korean Journal Of Food And Nutrition [issn:1225-4339 meta:br/0603]', 'volume': '25', 'issue': '1', 'page': '69-76', 'type': 'journal article', 'publisher': 'The Korean Society Of Food And Nutrition [crossref:4768 meta:ra/0604]', 'editor': 'Chung, Myong-Soo [orcid:0000-0002-9666-2513 meta:ra/0605]'}, 
-            {'id': 'doi:10.9799/ksfan.2012.25.1.077 meta:br/0602', 'title': 'Properties Of Immature Green Cherry Tomato Pickles', 'author': 'Koh, Jong-Ho [meta:ra/0606]; Shin, Hae-Hun [meta:ra/0607]; Kim, Young-Shik [orcid:0000-0001-5673-6314 meta:ra/0608]; Kook, Moo-Chang [meta:ra/0609]', 'pub_date': '2012-03-31', 'venue': 'The Korean Journal Of Food And Nutrition [issn:1225-4339 meta:br/0603]', 'volume': '', 'issue': '2', 'page': '77-82', 'type': 'journal article', 'publisher': 'The Korean Society Of Food And Nutrition [crossref:4768 meta:ra/0604]', 'editor': ''}]
         expected_output = [
             {'id': 'doi:10.17117/na.2015.08.1067 meta:br/0601', 'title': '', 'author': '', 'pub_date': '', 'venue': 'Scientometrics [issn:0138-9130 issn:1588-2861 meta:br/0603]', 'volume': '26', 'issue': '', 'page': '', 'type': 'journal article', 'publisher': 'Consulting Company Ucom [crossref:6623 meta:ra/0601]', 'editor': 'Naimi, Elmehdi [orcid:0000-0002-4126-8519 meta:ra/0602]'}, 
             {'id': 'issn:1524-4539 issn:0009-7322 meta:br/0602', 'title': 'Circulation', 'author': '', 'pub_date': '', 'venue': '', 'volume': '', 'issue': '', 'page': '', 'type': 'journal', 'publisher': '', 'editor': ''}, 
@@ -43,8 +44,9 @@ class test_ProcessTest(unittest.TestCase):
             {'id': 'doi:10.9799/ksfan.2012.25.1.077 meta:br/0606', 'title': 'Properties Of Immature Green Cherry Tomato Pickles', 'author': 'Koh, Jong-Ho [meta:ra/0608]; Shin, Hae-Hun [meta:ra/0609]; Kim, Young-Shik [orcid:0000-0001-5673-6314 meta:ra/06010]; Kook, Moo-Chang [meta:ra/06011]', 'pub_date': '2012-03-31', 'venue': 'The Korean Journal Of Food And Nutrition [issn:1225-4339 meta:br/0607]', 'volume': '', 'issue': '2', 'page': '77-82', 'type': 'journal article', 'publisher': 'The Korean Society Of Food And Nutrition [crossref:4768 meta:ra/0606]', 'editor': ''}]
         output = sorted(sorted(d.items()) for d in output)
         expected_output = sorted(sorted(d.items()) for d in expected_output)
+        self.maxDiff = None
         shutil.rmtree(output_folder)
-        delete_output_zip('.')
+        delete_output_zip('.', now)
         self.assertEqual(output, expected_output)
 
     def test_run_meta_process_two_workers(self):
@@ -52,13 +54,14 @@ class test_ProcessTest(unittest.TestCase):
         output_folder = os.path.join(BASE_DIR, 'output_2')
         meta_process = MetaProcess(config=os.path.join(BASE_DIR, 'meta_config_2.yaml'))
         meta_process.workers_number = 2
+        now = datetime.now()
         run_meta_process(meta_process)
         output = list()
         for dirpath, _, filenames in os.walk(os.path.join(output_folder, 'csv')):
             for file in filenames:
                 output.extend(get_data(os.path.join(dirpath, file)))
         shutil.rmtree(output_folder)
-        delete_output_zip('.')
+        delete_output_zip('.', now)
         expected_output = [
             {'id': 'doi:10.17117/na.2015.08.1067 meta:br/06101', 'title': '', 'author': '', 'pub_date': '', 'venue': 'Scientometrics [issn:0138-9130 issn:1588-2861 meta:br/06103]', 'volume': '26', 'issue': '', 'page': '', 'type': 'journal article', 'publisher': 'Consulting Company Ucom [crossref:6623 meta:ra/06101]', 'editor': 'Naimi, Elmehdi [orcid:0000-0002-4126-8519 meta:ra/06102]'}, 
             {'id': 'issn:1524-4539 issn:0009-7322 meta:br/06102', 'title': 'Circulation', 'author': '', 'pub_date': '', 'venue': '', 'volume': '', 'issue': '', 'page': '', 'type': 'journal', 'publisher': '', 'editor': ''}, 
@@ -73,6 +76,7 @@ class test_ProcessTest(unittest.TestCase):
         output_folder = os.path.join(BASE_DIR, 'output_3')	
         meta_process = MetaProcess(config=os.path.join(BASE_DIR, 'meta_config_3.yaml'))
         meta_process.input_csv_dir = os.path.join(BASE_DIR, 'input')
+        now = datetime.now()
         run_meta_process(meta_process)
         meta_process.input_csv_dir = os.path.join(BASE_DIR, 'input_2')
         run_meta_process(meta_process)
@@ -102,7 +106,6 @@ class test_ProcessTest(unittest.TestCase):
                         'https://w3id.org/oc/ontology/hasUpdateQuery': {
                             '<https://w3id.org/oc/meta/br/0601> <http://purl.org/spar/pro/isDocumentContextFor> <https://w3id.org/oc/meta/ar/0601> .', 
                             '<https://w3id.org/oc/meta/br/0601> <http://purl.org/vocab/frbr/core#partOf> <https://w3id.org/oc/meta/br/0603> .', 
-                            '<https://w3id.org/oc/meta/br/0601> <http://purl.org/spar/pro/isDocumentContextFor> <https://w3id.org/oc/meta/ar/0603> .', 
                             '<https://w3id.org/oc/meta/br/0601> <http://purl.org/spar/pro/isDocumentContextFor> <https://w3id.org/oc/meta/ar/0602> .', 
                             '<https://w3id.org/oc/meta/br/0601> <http://prismstandard.org/namespaces/basic/2.0/publicationDate> "2015-08-22"^^<http://www.w3.org/2001/XMLSchema#date> .', 
                             '<https://w3id.org/oc/meta/br/0601> <http://purl.org/dc/terms/title> "Some Aspects Of The Evolution Of Chernozems Under The Influence Of Natural And Anthropogenic Factors" .'}}]}, 
@@ -119,7 +122,8 @@ class test_ProcessTest(unittest.TestCase):
             're': [
                 {'@graph': [{'@id': 'https://w3id.org/oc/meta/re/0601/prov/se/1', 'http://www.w3.org/ns/prov#specializationOf': [{'@id': 'https://w3id.org/oc/meta/re/0601'}]}]}]}
         shutil.rmtree(output_folder)
-        delete_output_zip('.')
+        delete_output_zip('.', now)
+        self.maxDiff = None
         self.assertEqual(output, expected_output)
